backend/requirements.txt
flask==2.0.1
Pillow==9.0.0
opencv-python==4.5.3.56
numpy==1.21.2
requests==2.26.0
werkzeug==2.0.1
flask-cors==3.0.10
